wip : add Connection class - manage listening and client sockets - poll loop - maps sockets to servers
